move outdated/incompatible functions to resource
